multiple inheritance implemented in baseElement class
