+++ headers/thermalProperties.h
@@ -1,0 +1,36 @@
+#ifndef THERMALPROPERITES_H_INCLUDED
+#define THERMALPROPERITES_H_INCLUDED
+
+#include <iostream>
+using namespace std;
+
+class thermalProperties
+{
+protected:
+    double latentHeatOfFusion;
+    double latentHeatOfVaporization;
+    double specificHeatSolid;
+    double specificHeatLiquid;
+    double specificHeatGas;
+    double meltingPoint;
+    double boilingPoint;
+
+    void setLatentHeatOfFusion(double value) { latentHeatOfFusion = value; }
+    void setLatentHeatOfVaporization(double value) { latentHeatOfVaporization = value; }
+    void setSpecificHeatSolid(double value) { specificHeatSolid = value; }
+    void setSpecificHeatLiquid(double value) { specificHeatLiquid = value; }
+    void setSpecificHeatGas(double value) { specificHeatGas = value; }
+    void setMeltingPoint(double value) { meltingPoint = value; }
+    void setBoilingPoint(double value) { boilingPoint = value; }
+
+public:
+    double getLatentHeatOfFusion() { return latentHeatOfFusion; }
+    double getLatentHeatOfVaporization() { return latentHeatOfVaporization; }
+    double getSpecificHeatSolid() { return specificHeatSolid; }
+    double getSpecificHeatLiquid() { return specificHeatLiquid; }
+    double getSpecificHeatGas() { return specificHeatGas; }
+    double getMeltingPoint() { return meltingPoint; }
+    double getBoilingPoint() { return boilingPoint; }
+};
+
+#endif  // THERMALPROPERITES_H_INCLUDED
